adding main function and refactoring code to use separate packages
